requirements.txt
streamlit~=1.13.0
numpy~=1.23.4
requests~=2.28.1
Pillow~=9.2.0
keras~=2.11.0
beautifulsoup4~=4.11.1